requirements.txt
asgiref==3.5.2
Django==4.0.5
djangorestframework==3.13.1
mssql-django==1.1.3
mysqlclient==2.1.0
pyodbc==4.0.32
pytz==2022.1
sqlparse==0.4.2
tzdata==2022.1
